requirements.txt
Flask
psycopg2-binary
python-dotenv
PyPDF2
pytesseract
openai 